% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{check_committee_arg}
\alias{check_committee_arg}
\title{Check that a committee argument to a membership function returns only one
committee}
\usage{
check_committee_arg(committee)
}
\arguments{
\item{committee}{A vector of committee ids that should contain only one id.}
}
\description{
Check that a committee argument to a membership function returns only one
committee
}
\keyword{internal}
