% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{check_committee_id}
\alias{check_committee_id}
\title{Check that a committee argument to a membership function returns only one
committee}
\usage{
check_committee_id(committee_id)
}
\arguments{
\item{committee_id}{A single committee id.}
}
\description{
Check that a committee argument to a membership function returns only one
committee
}
\keyword{internal}
