% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{check_member_arg}
\alias{check_member_arg}
\title{Check that a member argument to a membership function returns only one
member}
\usage{
check_member_arg(member)
}
\arguments{
\item{committee}{A vector of mnis ids that should contain only one id.}
}
\description{
Check that a member argument to a membership function returns only one
member
}
\keyword{internal}
