% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{check_member_id}
\alias{check_member_id}
\title{Check that a member_id argument to a membership function returns only one
member}
\usage{
check_member_id(member_id)
}
\arguments{
\item{member_id}{A single mnis_id.}
}
\description{
Check that a member_id argument to a membership function returns only one
member
}
\keyword{internal}
