% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcommittees.R
\docType{package}
\name{clcommittees}
\alias{clcommittees}
\title{clcommittees: Tools for retreiving data from the Committees Information
System API}
\description{
The clcommittees package provides a suit of tools for downloading and
processing data from the UK Parliament's Committees Information System API.
}
