% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Cleans the column names of a data frame created from Committees API data}
\usage{
clean_column_names(df, remove_prefix)
}
\arguments{
\item{df}{A data frame with API variables column names.}

\item{reverse}{A common column name prefix which should be removed.}
}
\description{
Cleans the column names of a data frame created from Committees API data
}
\keyword{internal}
