% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{fetch_all_roles}
\alias{fetch_all_roles}
\title{Fetch data on the roles of the current and former members of a committee as
a tibble}
\usage{
fetch_all_roles(committee)
}
\arguments{
\item{committee}{An integer representing the id of the committee for which
to fetch the roles for the former members.}
}
\description{
\code{fetch_all_roles} fetches data on the roles of the current and former
members of a given committee and returns it as a tibble containing one row
per committee role.
}
\details{
A role indicates a period of service in a given position, so this function
returns ALL the roles for this committee (both current and historic) for its
current and former members. A role without an end date is a current role.

A member may have concurrent roles for the same period reflecting different
roles e.g. one indicating their service as a member and another their
service as a chair.
}
