% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees.R
\name{fetch_committee_types}
\alias{fetch_committee_types}
\title{Fetch data on the types of committees as a tibble}
\usage{
fetch_committee_types(committees = NULL)
}
\arguments{
\item{committees}{A vector of committee ids specifying the committees for
which to return chairs. The default is NULL, which returns data on the
chairs of all committees.}
}
\description{
\code{fetch_types} fetches data on the types of comittees and returns it as
a tibble containing one row per combination of committee and tyoe. The data
returned for each type is the data from the committee_types table of each
committee in the full table returned from \code{fetch_committees}.
}
\details{
You can optionally use the \code{committees} argument to provide a vector of
committee ids and the function will return just the types of the given
committees.
}
