% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees.R
\name{fetch_committees}
\alias{fetch_committees}
\title{Fetch data on all committees as a tibble}
\usage{
fetch_committees(current = FALSE, summary = TRUE)
}
\arguments{
\item{current}{A boolean indicating whether to return only the committees
without an end date. The default is FALSE.}

\item{summary}{A boolean indicating whether to return the summary columns or
the the full table. The default is TRUE.}
}
\description{
\code{fetch_committees} fetches data on all committees and returns it as a
tibble containing one row per committee.
}
\details{
By default this function returns a subset of the columns and ignores any
nested and redundant columns. Set \code{summary = FALSE} when calling the
function to retrieve the full data.
}
