% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees.R
\name{fetch_committees_for_url}
\alias{fetch_committees_for_url}
\title{Fetch data on current and former committees as a tibble using a Members
endpoint URL}
\usage{
fetch_committees_for_url(url, summary = TRUE)
}
\arguments{
\item{url}{A valid URL requesting data from the Committees endpoint.}

\item{summary}{A boolean indicating whether to exclude nested and empty
columns in the results. The default is TRUE.}
}
\description{
\code{fetch_committees_from_url} fetches data on committees from a Members
endpoint URL and returns it as a tibble containing one row per committee.
This internal function allows generic handling of requests for this data
to the Members endpoint with different URL parameters.
}
\keyword{internal}
