% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees.R
\name{fetch_current_chairs}
\alias{fetch_current_chairs}
\title{Fetch data on the current chairs of committees as a tibble}
\usage{
fetch_current_chairs(committees = NULL, summary = TRUE)
}
\arguments{
\item{committees}{A vector of committee ids specifying the committees for
which to return chairs. The default is NULL, which returns data on the
chairs of all committees.}

\item{summary}{A boolean indicating whether to return the summary columns or
the the full table. The default is TRUE.}
}
\description{
\code{fetch_chairs} fetches data on the current chairs of comittees and
returns it as a tibble containing one row per combination of committee and
chair. The data returned for each chair is the data from the chairs table of
each committee in the full table returned from \code{fetch_committees}.
}
\details{
You can optionally use the \code{committees} argument to provide a vector of
committee ids and the function will return just the current chairs of the
given committees.

By default this function returns a subset of the columns and ignores any
nested and redundant columns. Set \code{summary = FALSE} when calling the
function to retrieve the full data.
}
