% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees.R
\name{fetch_current_committees}
\alias{fetch_current_committees}
\title{Fetch data on current committees as a tibble}
\usage{
fetch_current_committees(summary = TRUE)
}
\arguments{
\item{summary}{A boolean indicating whether to return the summary columns or
the the full table. The default is TRUE.}
}
\description{
\code{fetch_current_committees} fetches data on current committees and
returns it as a tibble containing one row per committee. Current committees
are those that do not have an end date.
}
\details{
By default this function returns a subset of the columns and ignores any
nested and redundant columns. Set \code{summary = FALSE} when calling the
function to retrieve the full data.
}
