% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{fetch_current_members}
\alias{fetch_current_members}
\title{Fetch data on the current members of a committee as a tibble}
\usage{
fetch_current_members(committee, summary = TRUE)
}
\arguments{
\item{committee}{An integer representing the id of the committee for which
to fetch the current members.}

\item{summary}{A boolean indicating whether to return the summary columns or
the full table. The default is TRUE.}
}
\description{
\code{fetch_current_members} fetches data on the current members of a given
committee and returns it as a tibble containing one row per committee
member.
}
