% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{fetch_current_memberships}
\alias{fetch_current_memberships}
\title{Fetch data on the current members of a committee as a tibble}
\usage{
fetch_current_memberships(committee, summary = TRUE)
}
\arguments{
\item{committee}{An integer representing the id of the committee for which
to fetch the current memberships.}

\item{summary}{A boolean indicating whether to exclude nested and empty
columns in the results. The default is TRUE.}
}
\description{
\code{fetch_current_memberships} fetches data on the current members of a
given committee and returns it as a tibble containing one row per committee
membership.
}
