% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{fetch_current_roles}
\alias{fetch_current_roles}
\title{Fetch data on the roles of the current members of a committee as a tibble}
\usage{
fetch_current_roles(committee)
}
\arguments{
\item{committee}{An integer representing the id of the committee for which
to fetch the roles for the current members.}
}
\description{
\code{fetch_curent_roles} fetches data on the roles of the current members
of a given committee and returns it as a tibble containing one row per
committee role.
}
\details{
A role indicates a period of service in a given position, so this function
returns ALL the roles for this committee (both current and historic) for its
current members. A role without an end date is a current role.

A member may have concurrent roles for the same period reflecting different
positions e.g. one indicating their service as a member and another their
service as a chair.
}
