% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{fetch_former_memberships}
\alias{fetch_former_memberships}
\title{Fetch data on the former members of a committee as a tibble}
\usage{
fetch_former_memberships(committee_id, summary = TRUE, fetch_name = TRUE)
}
\arguments{
\item{committee_id}{An integer representing the id of the committee for
which to fetch the former memberships.}

\item{summary}{A boolean indicating whether to exclude nested and empty
columns in the results. The default is TRUE.}

\item{fetch_name}{A boolean indicating whether to add the committee name
by making an extra call to the API. The default is TRUE.}
}
\description{
\code{fetch_former_memberships} fetches data on the former members of a
given committee and returns it as a tibble containing one row per committee
membership.
}
