% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{fetch_memberships_for_member}
\alias{fetch_memberships_for_member}
\title{Fetch data on the current and former committee memberships of a given member}
\usage{
fetch_memberships_for_member(member, summary = TRUE)
}
\arguments{
\item{member}{An integer representing the mnis_id of the member for which
to fetch the current and former memberships.}

\item{summary}{A boolean indicating whether to return the summary columns or
the full table. The default is TRUE.}
}
\description{
\code{fetch_memberships_for_member} fetches data on the current and former
committee memberships of a given member and returns it as a tibble
containing one row per committee membership.
}
