% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{fetch_memberships_for_member_from_url}
\alias{fetch_memberships_for_member_from_url}
\title{Fetch data on the committee memberships of a given member as a tibble using
a Members endpoint URL}
\usage{
fetch_memberships_for_member_from_url(url, summary = TRUE)
}
\arguments{
\item{url}{A valid URL requesting data for the given member id from the
Members endpoint.}

\item{summary}{A boolean indicating whether to exclude nested and empty
columns in the results. The default is TRUE.}
}
\description{
\code{fetch_memberships_for_member_from_url} fetches data on the committee
memberships of a given member from a Members endpoint URL and returns it as
a tibble containing one row per committee membership. This internal function
allows generic handling of requests for this data to the Members endpoint
with different URL parameters.
}
