% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{fetch_memberships_from_url}
\alias{fetch_memberships_from_url}
\title{Fetch data on members of a given committee as a tibble using a Members
endpoint URL}
\usage{
fetch_memberships_from_url(
  committee_id,
  url,
  summary = TRUE,
  fetch_name = TRUE
)
}
\arguments{
\item{committee_id}{An integer representing the id of the committee for
which to fetch members.}

\item{url}{A valid URL requesting data for the given committee id from the
Memberships endpoint.}

\item{summary}{A boolean indicating whether to exclude nested and empty
columns in the results. The default is TRUE.}

\item{fetch_name}{A boolean indicating whether to add the committee name
by making an extra call to the API. The default is TRUE.}
}
\description{
\code{fetch_memberships_from_url} fetches data on the members of a given
committee from a Members endpoint URL and returns it as a tibble containing
one row per committee membership. This internal function allows generic
handling of requests for this data to the Members endpoint with different
URL parameters.
}
\keyword{internal}
