% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees.R
\name{fetch_scrutinising_departments}
\alias{fetch_scrutinising_departments}
\title{Fetch data on the departments scrutinised by committees as a tibble}
\usage{
fetch_scrutinising_departments(committees = NULL)
}
\arguments{
\item{committees}{A vector of committee ids specifying the committees for
which to return departments. The default is NULL, which returns data on
the departments scrutinised by all committees.}
}
\description{
\code{fetch_scrutinising_departments} fetches data on the government
departments scrutinised by comittees and returns it as a tibble containing
one row per combination of committee and department. The data returned for
each department is the data from the scrutinising_departments table of each
committee in the full table returned from \code{fetch_committees}.
}
\details{
You can optionally use the \code{committees} argument to provide a vector of
committee ids and the function will return just the departments of the given
committees.
}
