% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees.R
\name{fetch_sub_committees}
\alias{fetch_sub_committees}
\title{Fetch data on the subcommittees of parent committees as a tibble}
\usage{
fetch_sub_committees(committees = NULL)
}
\arguments{
\item{committees}{A vector of committee ids specifying the parent committees
for which to return any subcommittees. The default is NULL, which returns
data on subcommittees for all parent committees.}
}
\description{
\code{fetch_sub_committees} fetches data on the subcommittees of parent
comittees and returns it as a tibble containing one row per combination of
parent and subcommittee. The data returned for each subcommittee is the data
from the sub_committees table of its parent committee in the full table
returned from \code{fetch_committees}.
}
\details{
Any committee that is a subcommittee will also be listed as a committee in
the table returned from \code{fetch_committees}, so you may wish to use the
results of \code{fetch_sub_committees} to filter the results of
\code{fetch_committees}.

You can optionally use the \code{committees} argument to provide a vector of
committee ids and the function will return just the subcommittees of the
given parent committees.
}
