% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_ids}
\alias{format_ids}
\title{Formats a vector of numeric ids for inclusion as strings in an API URL}
\usage{
format_ids(ids)
}
\arguments{
\item{ids}{A vector of numeric ids to be formatted}
}
\description{
Formats a vector of numeric ids for inclusion as strings in an API URL
}
\keyword{internal}
