% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{process_memberships}
\alias{process_memberships}
\title{Processes the data returned from a request to the membership endpoints.}
\usage{
process_memberships(memberships, summary = TRUE)
}
\arguments{
\item{memberships}{A tibble returned from \code{request_items} called with
the membership/member endpoint as the url.}

\item{summary}{A boolean indicating whether to return the summary columns or
the full table. The default is TRUE.}
}
\description{
Processes the data returned from a request to the membership endpoints.
}
\keyword{internal}
