% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{process_roles}
\alias{process_roles}
\title{Processes the full table returned from \code{fetch_members} or
\code{fetch_memberships_for} to exract the data on roles.}
\usage{
process_roles(roles)
}
\arguments{
\item{roles}{A tibble returned from \code{fetch_members} or
\code{fetch_memberships_for} called with \code{summary = FALSE}.}
}
\description{
Processes the full table returned from \code{fetch_members} or
\code{fetch_memberships_for} to exract the data on roles.
}
