% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_column_prefix}
\alias{remove_column_prefix}
\title{Removes a repeating prefix from the column names of a data frame}
\usage{
remove_column_prefix(df, remove_prefix)
}
\arguments{
\item{df}{A data frame with column names that need cleaning.}

\item{reverse}{A common column name prefix which should be removed.}
}
\description{
Removes a repeating prefix from the column names of a data frame
}
\keyword{internal}
