% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{request}
\alias{request}
\title{Send a request to API and return the response as a named list}
\usage{
request(url)
}
\arguments{
\item{url}{The full API URL specifying the endpoint and request parameters.}
}
\description{
Send a request to API and return the response as a named list
}
