% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{request_items}
\alias{request_items}
\title{Send a request to API and return the response items as a tibble}
\usage{
request_items(url)
}
\arguments{
\item{url}{The full API URL specifying the endpoint and request parameters.}
}
\description{
\code{request_items} makes an API call to the given endpoint, extracts
the data items from the response as a tibble, and cleans the column names.
This function is intended to be used with API requests that ask for all
the available items with \code{?parameters.all=true}. Do not use this
function if you need to access any of the other properties of the response
object; call \code{request} and \code{get_response_items} separately.
}
