% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memberships.R
\name{process_memberships}
\alias{process_memberships}
\title{Processes a tibble returned from \code{request_items} when called with the
url of either the member or membership endpoint.}
\usage{
process_memberships(memberships, summary = TRUE)
}
\arguments{
\item{memberships}{A tibble returned from \code{request_items} called with
the membership/member endpoint as the url.}

\item{summary}{A boolean indicating whether to return the summary columns or
the full table. The default is TRUE.}
}
\description{
Processes a tibble returned from \code{request_items} when called with the
url of either the member or membership endpoint.
}
\keyword{internal}
